+++ jest.config.js
@@ -3,7 +3,7 @@ const { compilerOptions } = require("./tsconfig.json");
 /** @type {import('jest').Config} */
 const config = {
   coverageReporters: ["json-summary", "lcov", "text-summary"],
-  coverageThreshold: { global: { branches: 32, functions: 54, lines: 50, statements: 52 } },
+  coverageThreshold: { global: { branches: 38, functions: 61, lines: 57, statements: 58 } },
   detectOpenHandles: true,
   forceExit: true,
   modulePaths: [compilerOptions.baseUrl],