+++ src/modules/cleanup/cleanup.ts
@@ -62,7 +62,6 @@ export class Cleanup {
    * Removes unused users
    */
   private async cleanupUsers(): Promise<void> {
-    const monthAgoDate = new Date(Date.now() - 30 * 24 * 60 * 60 * 1000);
     // Remove unused users from database
     await this.database.user.deleteMany({
       where: {
@@ -72,6 +71,8 @@ export class Cleanup {
           { chatEditors: { none: {} } },
           { chatSettingsHistoryAuthors: { none: {} } },
           { chatSettingsHistoryEditors: { none: {} } },
+          { messageAuthors: { none: {} } },
+          { messageEditors: { none: {} } },
           { profaneWordAuthors: { none: {} } },
           { profaneWordEditors: { none: {} } },
           { senderChatAuthors: { none: {} } },
@@ -93,12 +94,13 @@ export class Cleanup {
       select: { id: true, userAuthors: { select: { id: true } }, userEditors: { select: { id: true } } },
       where: {
         AND: [
-          { updatedAt: { lt: monthAgoDate } },
           { chatAdmins: { none: {} } },
           { chatAuthors: { none: {} } },
           { chatEditors: { none: {} } },
           { chatSettingsHistoryAuthors: { none: {} } },
           { chatSettingsHistoryEditors: { none: {} } },
+          { messageAuthors: { none: {} } },
+          { messageEditors: { none: {} } },
           { profaneWordAuthors: { none: {} } },
           { profaneWordEditors: { none: {} } },
           { senderChatAuthors: { none: {} } },