+++ src/test/mockUser.ts
@@ -14,3 +14,11 @@ export const mockUser = (user?: Partial<User>): User => ({
   username: "john_doe",
   ...user,
 });
+
+/**
+ * Mocks the second Telegram user
+ * @param user Telegram user which will be merged with the default one
+ * @returns Telegram user
+ */
+export const mockUser2 = (user?: Partial<User>): User =>
+  mockUser({ first_name: "Jack", id: 2, last_name: "Green", username: "jack_green", ...user });