+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.13.0 (2024-xx-xx)
 
 - Improvement: Messages On Behalf Of Channels feature
+- Improvement: based on the results of voting for a ban, related messages will be deleted as well (if there are several images in the message)
 - Improvement: updated tooltip for using the voteban command
 - Other: internal improvements